Fix check for missing an MIT email
I had a core problem with the logic here -- you must *also* be trying to
claim MIT student status in order to be regarded as missing a verified
email. This was breaking the dues payment form.

Let's add some missing test coverage, and also add `ng-cloak` to prevent
temporary rendering of warning messages.